requirements.txt
python-dotenv=1.0.1
interactions.py=5.13.2
CairoSVG=2.7.1